+++ irc/capability.go
@@ -18,6 +18,7 @@ const (
 	AccountNotify   Capability = "account-notify"
 	AwayNotify      Capability = "away-notify"
 	CapNotify       Capability = "cap-notify"
+	EchoMessage     Capability = "echo-message"
 	ExtendedJoin    Capability = "extended-join"
 	InviteNotify    Capability = "invite-notify"
 	MessageTags     Capability = "draft/message-tags"
@@ -33,6 +34,7 @@ var (
 		AccountNotify: true,
 		AwayNotify:    true,
 		CapNotify:     true,
+		EchoMessage:   true,
 		ExtendedJoin:  true,
 		InviteNotify:  true,
 		MessageTags:   true,