+++ CHANGELOG.md
@@ -13,7 +13,7 @@ New release of Oragono!
 ### Added
 * Added `REHASH` command.
 * Added ability to enable and disable SASL.
-* Added support for IRCv3 capability [`cap-notify`](http://ircv3.net/specs/extensions/cap-notify-3.2.html).
+* Added support for IRCv3 capabilities [`cap-notify`](http://ircv3.net/specs/extensions/cap-notify-3.2.html) and [`echo-message`](http://ircv3.net/specs/extensions/echo-message-3.2.html).
 
 ### Changed